first stage of routing done
