+++ client/src/App.tsx
@@ -1,31 +1,29 @@
 import React from 'react';
-import Header from './views/Header';
-import { CssBaseline } from '@material-ui/core';
-import Container from '@material-ui/core/Container';
-import Main from './views/Main';
+import Home from './views/Home';
 import axios from 'axios';
-
-let sections: {title: string, url: string} [] = [
-  {title : "Login" , url : "#"},
-  {title : "Clients", url : "#"},
-  {title : "About", url : "#"}
-]
+import {
+  BrowserRouter as Router,
+  Switch,
+  Route,
+  Link
+} from "react-router-dom";
+import Login from './views/Login';
 
 axios.defaults.baseURL = process.env.REACT_APP_API_URL;
 
 export default function App() {
   
   return (
-    <React.Fragment>
-      <CssBaseline>
-      <Container maxWidth="lg">
-        <Header title="PK" sections={sections}></Header>
-        <main>
-          <Main></Main>
-        </main>
-    </Container>
-    </CssBaseline>
-    </React.Fragment>
+    <Router>
+      <Switch>
+        <Route exact path = "/">
+          <Home></Home>
+        </Route>
+        <Route path="/login">
+          <Login></Login>
+        </Route>
+      </Switch>
+    </Router>
   );
 }
