+++ client/src/views/Login.tsx
@@ -18,7 +18,7 @@ function Copyright() {
   return (
     <Typography variant="body2" color="textSecondary" align="center">
       {'Copyright © '}
-      <Link color="inherit" href="#">
+      <Link color="inherit" href="/">
         PK
       </Link>{' '}
       {new Date().getFullYear()}
